+++ src/main/java/ru/intel/credits/repository/RecieveDBRepository.java
@@ -4,6 +4,7 @@ import ru.intel.credits.model.*;
 
 import java.util.Collection;
 import java.util.HashMap;
+import java.util.concurrent.atomic.AtomicLong;
 
 public interface RecieveDBRepository {
 
@@ -11,5 +12,5 @@ public interface RecieveDBRepository {
 
     void insertAllDebts(Collection<Debt> debts, HashMap<Long, VidDebt> dirDebts);
 
-    Integer getSequence();
+    AtomicLong getSequence();
 }
